add payment new

+++ BusinessLogic/Models/Momo/MomoExecuteResponseModel.cs
@@ -11,7 +11,6 @@ namespace BusinessLogic.Models.Momo
 		public string OrderId { get; set; }
 		public string Amount { get; set; }
 		public string OrderInfo { get; set; }
-		public long CourseId { get; set; }
-
+		public int CourseId { get; set; }
 	}
 }
